rename menuBar add Mastodon link
